Refactor CardPost styles
