+++ components/Cards/CardPost.jsx
@@ -1,5 +1,6 @@
-import { Image, StyleSheet, Text, View } from "react-native";
+import { Image, Text, View } from "react-native";
 import MaterialCommunityIcons from "@expo/vector-icons/MaterialCommunityIcons";
+import styles from "../../styles/CardPost";
 
 export default function CardPost({ userName, date, body, likeCount, commentCount }) {
   return (
@@ -43,45 +44,3 @@ export default function CardPost({ userName, date, body, likeCount, commentCount
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  userImage: {
-    width: 38,
-    height: 38,
-    borderRadius: 19,
-  },
-  cardContainer: {
-    width: 335,
-    height: 135,
-    paddingTop: 10,
-    paddingLeft: 10,
-    borderRadius: 15,
-    borderWidth: 2,
-    borderColor: "#295C53",
-    flexDirection: "row",
-    gap: 12,
-  },
-  postInfo: {
-    flexDirection: "row",
-    gap: 25,
-  },
-  userName: {
-    fontSize: 16,
-    fontWeight: "700",
-    color: "#223F3A",
-  },
-  postDate: {
-    fontSize: 12,
-    fontWeight: "300",
-    color: "#6D6D6D",
-  },
-  postBody: {
-    fontSize: 14,
-    fontWeight: "400",
-    maxWidth: 250,
-  },
-  postInfoText: {
-    fontSize: 10,
-    marginTop: 10,
-  },
-});
